If the bot doesn't have the permissions needed to delete a maze channel, it doesn't try anymore

+++ events/ready.js
@@ -130,6 +130,9 @@ module.exports = async bot => {
 
 			if ((Date.now() - message.createdTimestamp) < 300000) return;
 
+			const permissions = channel.permissionsFor(channel.guild.me);
+			if (!permissions.has("MANAGE_CHANNELS")) return;
+
 			channel.delete().catch(e => bot.logger.log("error", e.stack));
 		});
